Updated README and workaround build problem
Updated README.md to reflect the current way of building.

Removed "upcalls" so that the resulting libmmtk_ruby.so does not depend
on symbols in the Ruby VM.  This fixes some build problems related to
circular dependencies between `ruby` and `mmtk-ruby`.  It is planned to
support upcalls by letting the Ruby VM register callback functions to
mmtk-ruby.