Temporarily depend on wks/mmtk-core
There will be a change in the mmtk-core repository that changes the
root-scanning API.  The hacks that implemented object-enqueuing tracing
in this binding will no longer work because the core no longer exposes
the ProcessEdgesWork type to the binding.

We use the following mmtk-core repository to develop the changes for
making Ruby working again:

https://github.com/wks/mmtk-core/tree/ruby-friendly-tracing

Once the changes are merged back to the official repo, we will let the
mmtk-ruby binding depend on the official repo again.