Update Ruby repo revision
The upstream made some changes to the root-scanning functions, notably,

-   Introduced yjit roots.
-   Seperated root-scanning into the GC-implementation-agnostic
    rb_gc_mark_roots in gc.c and the GC-implementation-specific
    mark_roots in gc/default.c.

In the `ruby` repo, we reorder the root-scanning functions in the upcall
table so that the general functions come before the
implementation-specific functions.  Also added yjit roots as one upcall,
although its implementation is a no-op if yjit is not enabled.