Ruby: Cherry-pick #80
The Ruby repository cherry-picked https://github.com/mmtk/ruby/pull/80
onto the dev/mmtk-overrides-default branch which will follow the
official master branch.  It added a CI scripts and excluded some tests
using Ruby scripts in the `.excludes-mmtk/` directory.

We add an option to `make test-all` to take the `.excludes-mmtk`
directory into consideration.

With this change, it should be, in theory, possible to selectively run
tests runnable for MMTk without using the whitelist
`ruby-test-cases.txt` in this repo.  However, some test cases still
crashes due to bugs in the binding.  We'll still use the whitelist for
running binding tests for the mmtk-core repo, until we fix those crashes
so that `make test-all` can finish safely with MMTk enabled.