Update rev, remove whitelist and change default plan
Updated the `ruby` repo.  It enabled more test cases that were disabled
before.

Removed the whitelist.  Now we rely on the exclusion list in
`test/.excludes-mmtk` in the `ruby` repo.

Changed the default plan to StickyImmix since it has been stable so far.

We also run bootstraptest/test_thread.rb with a fixed heap size in debug
build because it allocates too aggressively and may cause timeout in CI.