Unregister PPPs when they are no longer PPPs
This commit, together with a commit in the `ruby` repo, allows the GC to
remove objects from the list of registered PPPs when they are no longer
PPPs.  Specifically, iseq objects will stop being PPPs since
ISEQ_COMPILE_DATA_CLEAR.

This commit also adds eBPF tracing extensions to track the number of
PPPs and their children involved when pinning or unpinning PPP children,
and the number of PPPs removed when cleaning up the PPP list, and
visualize them on the generated timeline plot.