Bump Ruby repo revision and deps version
The ruby repo is merged with the upstream master.  The modular GC has
been introduced upstream, and an experimental MMTk binding as a GC
module has been included in the official repo.

The ruby repo also fixed missing implementation of `memsize` and
`imemo_type_ids` for MMTk-specific types, excluded a few tests that
assume the behavior of the default GC, and implemented `GC.count`.