Update ruby repo revision
The upstream changed its way of updating GIVTBL.  Now it no longer
updates generic ivars in gc_update_object_reference, but instead do so
for all entries in generic_iv_tbl_ in bulk during weak table updating
phase.  We still stick to the old method.  In the
dev/mmtk-overrides-default branch of the mmtk/ruby repo, we will still
use the old way, i.e. tracing and forwarding generic ivars when scanning
an object, when using MMTk.  No change in the Rust part of the code
(i.e. this repo) is needed.