Handle cc_refinement table and reorgainze code
Previously we didn't process the newly added cc_refinement table, and
caused crashes.

We re-organized the order of weak table processing functions, and follow
the order as listed in `rb_gc_vm_weak_table_foreach` in `gc.c`, so that
we don't miss a table.  We also replaced functions that get tables with
functions that simply get table sizes.