Update ruby and mmtk-core repos revisions
The `ruby` repo is merged with `master` on 2025-08-25.  The upstream is
now using `rb_gc_mark_weak` for more fields.  In response to that, we
properly implemented the handling of weak fields when using MMTk.
However, we are now temporarily treating WeakMap as a strong map because
it needs to be rehashed during the weak reference processing phase.  We
also added eBPF probes for weak fields.